refactor: remove clam-av scanner as actual scanning is done by submission-scanner-processor
