+++ .circleci/config.yml
@@ -64,6 +64,7 @@ workflows:
             branches:
               only:
                 - develop
+                - feat/cleanup-scanner
       - "build-prod":
           context: org-global
           filters:
