+++ README.md
@@ -1,17 +1,14 @@
-# Topcoder - Submission ClamAV API
+# Topcoder - Submission Scan API
 
 ## Requirements
 
 - NodeJS v16.+ (local deployment)
-- ClamAV (local deployment
 - Docker (docker deployment)
 
 ## Configuration
 
 The configurations can be changed in `config/default.js` (default environment) and `config/test.js` (test environment), or by setting OS environment variables. Here are the important configurations that need to configure correctly
 
-- `CLAMAV_HOST` the ClamAV daemon host (`TCPAddr` in `clam.conf`)
-- `CLAMAV_PORT` the ClamAV daemon port (`TCPSocket` in `clam.conf`)
 - `BUSAPI_EVENTS_URL` Bus API URL
 - `AVSCAN_TOPIC` AV Scan Kafka topic
 - All environment variables starting with `AUTH0` prefix
@@ -31,46 +28,6 @@ The configurations can be changed in `config/default.js` (default environment) a
 - Run `docker-compose build` to build the images (including the clamav-scan-api and ClamAV)
 - Run `docker-compose up` to run the app
 - The app will be available at `http://<your_docker_machine_ip>:3000/api/v1` by default
-  **NOTE** ClamAV deamon needs some time to get started (about 30 seconds in my environment), so wait for it before making calls to the API or running tests
 - You can run the tests by `docker exec -w /usr/src/app clamav-scan-api npm test`
 - You can run the lint by `docker exec -w /usr/src/app clamav-scan-api npm run lint`
 - You can run the lint:fix by `docker exec -w /usr/src/app clamav-scan-api npm run lint:fix`
-
-## Manual Verification
-
-- You can use `Postman` or `cURL` to make call to `http://localhost:3000/api/v1/scan` to verify
-- Example call with `cURL` on my Windows:
-
-```bash
-> curl -X POST http://localhost:3000/api/v1/scan -F "file=@C:\EICAR_submission.zip"
-{"infected":true,"malicious":"Win.Test.EICAR_HDB-1"}
-```
-
-## How to use the REST API
-
-- The API accepts POST request with content type `multipart/form-data`
-- The form field name should be `file`
-- The response json format includes 2 properties:
-
-```json
-{ "infected": true, "malicious": "Win.Test.EICAR_HDB-1" }
-```
-
-- `infected`: true if the file was infected by a malicious, otherwise false
-- `malicious`: (optional) the malicious found by ClamAV, only present if `infected` is true
-
-## Setup ClamAV
-
-- Go to https://www.clamav.net/downloads
-- Under `Alternate Versions of ClamAV` section, choose your OS and follow the instructions to download or install ClamAV (including `clamd`)
-- Copy `<ClamAV-directory>/conf_examples/clamd.conf.sample` to `<ClamAV-directory>/clamd.conf`
-- Copy `<ClamAV-directory>/conf_examples/freshclam.conf.sample` to `<ClamAV-directory>/freshclam.conf`
-- Comment-out line 8 in both `clamd.conf` and `freshclam.conf` files:
-
-```bash
-# Comment or remove the line below.
-# Example
-```
-
-- Run `freshclam` to update virus database
-- Run `clamd` to start ClamAV deamon
